Annotations in separate module
