+++ build.gradle
@@ -73,6 +73,50 @@ void setupMultiplatformLibrary(Project project) {
     project.apply from: '../publication.gradle'
 }
 
+void setupAllTargetsWithDefaultSourceSets(Project project) {
+    final target = project.currentTarget()
+
+    if (target.isCommon() || target.isMeta()) {
+        setupCommonTargets(project)
+
+        project.kotlin {
+            sourceSets {
+                jvmMain.dependsOn commonMain
+                jvmTest.dependsOn commonTest
+
+                androidMain.dependsOn commonMain
+                androidTest.dependsOn commonTest
+
+                jsMain.dependsOn commonMain
+                jsTest.dependsOn commonTest
+
+                linuxX64Main.dependsOn commonMain
+                linuxX64Test.dependsOn commonTest
+
+                linuxArm32HfpMain.dependsOn commonMain
+                linuxArm32HfpTest.dependsOn commonTest
+            }
+        }
+    }
+
+    if (target.isIos() || target.isMeta()) {
+        setupIosTargets(project)
+
+        project.kotlin {
+            sourceSets {
+                ios32Main.dependsOn commonMain
+                ios32Test.dependsOn commonTest
+
+                ios64Main.dependsOn commonMain
+                ios64Test.dependsOn commonTest
+
+                iosSimMain.dependsOn commonMain
+                iosSimTest.dependsOn commonTest
+            }
+        }
+    }
+}
+
 void setupAndroidTarget(Project project) {
     project.apply plugin: 'com.android.library'
 
@@ -94,14 +138,14 @@ void setupAndroidTarget(Project project) {
                     implementation kotlin('stdlib')
                 }
             }
-            
+
             androidTest {
                 dependencies {
                     implementation kotlin('test-junit')
                 }
             }
         }
-        
+
         android {
             publishLibraryVariants('release', 'debug')
         }
@@ -118,7 +162,7 @@ void setupJvmTarget(Project project) {
                     implementation kotlin('stdlib')
                 }
             }
-            
+
             jvmTest {
                 dependencies {
                     implementation kotlin('test-junit')
