+++ reaktive-annotations/src/commonMain/kotlin/com/badoo/reaktive/annotations/Annotations.kt
@@ -1,0 +1,43 @@
+package com.badoo.reaktive.annotations
+
+/**
+ * Indicates that return value should normally be used
+ */
+@Retention(AnnotationRetention.SOURCE)
+@Target(AnnotationTarget.FUNCTION)
+annotation class UseReturnValue
+
+/**
+ * Indicates that annotated source does not operate on any particular scheduler
+ */
+@Retention(AnnotationRetention.SOURCE)
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
+annotation class EventsOnUnspecifiedScheduler
+
+/**
+ * Indicates that events can be produced on any scheduler
+ */
+@Retention(AnnotationRetention.SOURCE)
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
+annotation class EventsOnAnyScheduler
+
+/**
+ * Indicates that all events are produced on Main scheduler
+ */
+@Retention(AnnotationRetention.SOURCE)
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
+annotation class EventsOnMainScheduler
+
+/**
+ * Indicates that all events are produced on Computation scheduler
+ */
+@Retention(AnnotationRetention.SOURCE)
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
+annotation class EventsOnComputationScheduler
+
+/**
+ * Indicates that all events are produced on IO scheduler
+ */
+@Retention(AnnotationRetention.SOURCE)
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
+annotation class EventsOnIoScheduler
