+++ reaktive-test/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.github.badoo.reaktive.test" />
+<manifest package="com.badoo.reaktive.test" />
