+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/maybe/Susbcribe.kt
@@ -1,9 +1,9 @@
 package com.badoo.reaktive.maybe
 
+import com.badoo.reaktive.annotations.UseReturnValue
 import com.badoo.reaktive.base.subscribeSafe
 import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
-import com.badoo.reaktive.utils.UseReturnValue
 import com.badoo.reaktive.utils.handleSourceError
 
 @UseReturnValue
