--- reaktive/src/commonMain/kotlin/com/badoo/reaktive/utils/Annotations.kt
@@ -1,8 +1,0 @@
-package com.badoo.reaktive.utils
-
-/**
- * Indicates that return value should be used
- */
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.FUNCTION)
-annotation class UseReturnValue
