+++ reaktive/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.github.badoo.reaktive" />
+<manifest package="com.badoo.reaktive" />
